Move PropertyValidationUtils to validation package
