+++ java/configuration/src/main/java/sleeper/configuration/properties/instance/AsyncIngestPartitionFileWriterProperty.java
@@ -16,8 +16,8 @@
 
 package sleeper.configuration.properties.instance;
 
-import sleeper.configuration.PropertyValidationUtils;
 import sleeper.configuration.properties.SleeperPropertyIndex;
+import sleeper.configuration.properties.validation.PropertyValidationUtils;
 
 import java.util.List;
