+++ java/configuration/src/test/java/sleeper/configuration/PropertyValidationUtilsTest.java
@@ -20,6 +20,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import sleeper.configuration.properties.validation.PropertyValidationUtils;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class PropertyValidationUtilsTest {
